Windows build fixes.
Windows doesn't have ssize_t, sadly. There's SSIZE_T, but defining an
OPENSSL_SSIZE_T seems worse than just using an int.

+++ include/openssl/base64.h
@@ -129,8 +129,7 @@ int EVP_DecodeFinal(EVP_ENCODE_CTX *ctx, uint8_t *out, int *out_len);
  *
  * WARNING: EVP_DecodeBlock's return value does not take padding into
  * account. TODO(davidben): Possible or worth it to fix or add new API? */
-ssize_t EVP_DecodeBlock(uint8_t *dst, const uint8_t *src, size_t src_len);
-
+int EVP_DecodeBlock(uint8_t *dst, const uint8_t *src, size_t src_len);
 
 struct evp_encode_ctx_st {
   unsigned num;    /* number saved in a partial encode/decode */
